requirements.txt
streamlit
python-dotenv
pandas
google-generativeai
matplotlib
streamlit-authenticator
seaborn
plotly
transformers
torch

